test(spec): rename _step0_parser to step0_parser (taste-lint naming)
The taste-lints check rejected the leading-underscore filename.
Renamed to drop the underscore; updated the import in test_spec_step0.py.

50 tests still pass.
